+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/exception/ActionNotFoundException.java
@@ -1,4 +1,4 @@
-package org.motechproject.tasks.ex;
+package org.motechproject.tasks.exception;
 
 /**
  * Thrown when the requested action doesn't exists.
