+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/exception/TaskNotFoundException.java
@@ -1,4 +1,4 @@
-package org.motechproject.tasks.ex;
+package org.motechproject.tasks.exception;
 
 import static java.lang.String.format;
