+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/osgi/TasksBlueprintApplicationContextTracker.java
@@ -2,7 +2,7 @@ package org.motechproject.tasks.osgi;
 
 import org.motechproject.osgi.web.tracker.ApplicationContextTracker;
 import org.motechproject.tasks.annotations.TaskAnnotationBeanPostProcessor;
-import org.motechproject.tasks.ex.ValidationException;
+import org.motechproject.tasks.exception.ValidationException;
 import org.motechproject.tasks.service.ChannelService;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
