+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/service/MethodHandler.java
@@ -3,7 +3,7 @@ package org.motechproject.tasks.service;
 import org.motechproject.tasks.domain.ActionEvent;
 import org.motechproject.tasks.domain.ActionParameter;
 import org.motechproject.tasks.domain.MethodCallManner;
-import org.motechproject.tasks.ex.TaskHandlerException;
+import org.motechproject.tasks.exception.TaskHandlerException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
