+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/service/TaskActionExecutor.java
@@ -10,8 +10,8 @@ import org.motechproject.tasks.domain.KeyInformation;
 import org.motechproject.tasks.domain.ParameterType;
 import org.motechproject.tasks.domain.Task;
 import org.motechproject.tasks.domain.TaskActionInformation;
-import org.motechproject.tasks.ex.ActionNotFoundException;
-import org.motechproject.tasks.ex.TaskHandlerException;
+import org.motechproject.tasks.exception.ActionNotFoundException;
+import org.motechproject.tasks.exception.TaskHandlerException;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceReference;
 import org.slf4j.Logger;
