+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/service/TaskContext.java
@@ -4,7 +4,7 @@ import org.apache.commons.beanutils.PropertyUtils;
 import org.motechproject.commons.api.MotechException;
 import org.motechproject.tasks.domain.Task;
 import org.motechproject.tasks.events.constants.TaskFailureCause;
-import org.motechproject.tasks.ex.TaskHandlerException;
+import org.motechproject.tasks.exception.TaskHandlerException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
