+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/service/impl/TaskServiceImpl.java
@@ -27,11 +27,11 @@ import org.motechproject.tasks.domain.TaskDataProvider;
 import org.motechproject.tasks.domain.TaskError;
 import org.motechproject.tasks.domain.TaskTriggerInformation;
 import org.motechproject.tasks.domain.TriggerEvent;
-import org.motechproject.tasks.ex.ActionNotFoundException;
-import org.motechproject.tasks.ex.CustomParserNotFoundException;
-import org.motechproject.tasks.ex.TaskNameAlreadyExistsException;
-import org.motechproject.tasks.ex.TaskNotFoundException;
-import org.motechproject.tasks.ex.ValidationException;
+import org.motechproject.tasks.exception.ActionNotFoundException;
+import org.motechproject.tasks.exception.CustomParserNotFoundException;
+import org.motechproject.tasks.exception.TaskNameAlreadyExistsException;
+import org.motechproject.tasks.exception.TaskNotFoundException;
+import org.motechproject.tasks.exception.ValidationException;
 import org.motechproject.tasks.repository.TasksDataService;
 import org.motechproject.tasks.service.ChannelService;
 import org.motechproject.tasks.service.TaskDataProviderService;
