+++ modules/tasks/tasks/src/test/java/org/motechproject/tasks/service/TaskContextTest.java
@@ -15,7 +15,7 @@ import org.motechproject.tasks.domain.Task;
 import org.motechproject.tasks.domain.TaskActionInformation;
 import org.motechproject.tasks.domain.TaskBuilder;
 import org.motechproject.tasks.events.constants.TaskFailureCause;
-import org.motechproject.tasks.ex.TaskHandlerException;
+import org.motechproject.tasks.exception.TaskHandlerException;
 
 import java.util.HashMap;
 import java.util.Map;
