+++ modules/tasks/tasks/src/test/java/org/motechproject/tasks/service/impl/TaskServiceImplTest.java
@@ -40,10 +40,10 @@ import org.motechproject.tasks.domain.TaskEvent;
 import org.motechproject.tasks.domain.TaskEventInformation;
 import org.motechproject.tasks.domain.TaskTriggerInformation;
 import org.motechproject.tasks.domain.TriggerEvent;
-import org.motechproject.tasks.ex.ActionNotFoundException;
-import org.motechproject.tasks.ex.TaskNameAlreadyExistsException;
-import org.motechproject.tasks.ex.TaskNotFoundException;
-import org.motechproject.tasks.ex.ValidationException;
+import org.motechproject.tasks.exception.ActionNotFoundException;
+import org.motechproject.tasks.exception.TaskNameAlreadyExistsException;
+import org.motechproject.tasks.exception.TaskNotFoundException;
+import org.motechproject.tasks.exception.ValidationException;
 import org.motechproject.tasks.repository.TasksDataService;
 import org.motechproject.tasks.service.ChannelService;
 import org.motechproject.tasks.service.TaskDataProviderService;
